[refactor] Rename Jira class module s/main/jira
