+++ test/cli/test_main_smoketest.py
@@ -11,7 +11,7 @@ import pytest
 
 from fixtures import ISSUE_1
 from jira_offline.cli import cli
-from jira_offline.main import Issue
+from jira_offline.jira import Issue
 
 
 # 0: CLI command name
